// libs/features/collection/billing/src/lib/shared/models/debit-credit-details.ts
export class DebitCreditDetails {
  CLOSE_CON: number = undefined;
  CLOSE_CON_ANN: number = undefined;
  CLOSE_CON_OH: number = undefined;
  CLOSE_CON_UI: number = undefined;
  CLOSE_CREDIT: number = undefined;
  CLOSE_PEN: number = undefined;
  CLOSE_PEN_ANN: number = undefined;
  CLOSE_PEN_OH: number = undefined;
  CLOSE_PEN_UI: number = undefined;
  CLOSING_DEBIT: number = undefined;
  CURRNT_ADJ: number = undefined;
  CURRNT_ADJPEN: number = undefined;
  CURRNT_ADJPEN_ALLOC: number = undefined;
  CURRNT_ADJPEN_ANN: number = undefined;
  CURRNT_ADJPEN_ANN_ALLOC: number = undefined;
  CURRNT_ADJPEN_OH: number = undefined;
  CURRNT_ADJPEN_OH_ALLOC: number = undefined;
  CURRNT_ADJPEN_UI: number = undefined;
  CURRNT_ADJPEN_UI_ALLOC: number = undefined;
  CURRNT_ADJ_ALLOC: number = undefined;
  CURRNT_ADJ_ANN: number = undefined;
  CURRNT_ADJ_ANN_ALLOC: number = undefined;
  CURRNT_ADJ_OH: number = undefined;
  CURRNT_ADJ_OH_ALLOC: number = undefined;
  CURRNT_ADJ_UI: number = undefined;
  CURRNT_ADJ_UI_ALLOC: number = undefined;
  CURRNT_CON: number = undefined;
  CURRNT_CON_ALLOC: number = undefined;
  CURRNT_CON_ANN: number = undefined;
  CURRNT_CON_ANN_ALLOC: number = undefined;
  CURRNT_CON_OH: number = undefined;
  CURRNT_CON_OH_ALLOC: number = undefined;
  CURRNT_CON_UI: number = undefined;
  CURRNT_CON_UI_ALLOC: number = undefined;
  CURRNT_PEN: number = undefined;
  CURRNT_PEN_ALLOC: number = undefined;
  CURRNT_PEN_ANN: number = undefined;
  CURRNT_PEN_ANN_ALLOC: number = undefined;
  CURRNT_PEN_OH: number = undefined;
  CURRNT_PEN_OH_ALLOC: number = undefined;
  CURRNT_PEN_UI: number = undefined;
  CURRNT_PEN_UI_ALLOC: number = undefined;
  ESTACCMAPPINGID: number = undefined;
  INJURY_DEDUCTION_REFUND_CREDIT: number = undefined;
  OPENINGBALANCECREDIT: number = undefined;
  OPEN_CON: number = undefined;
  OPEN_CON_ALLOC: number = undefined;
  OPEN_CON_ANN: number = undefined;
  OPEN_CON_ANN_ALLOC: number = undefined;
  OPEN_CON_OH: number = undefined;
  OPEN_CON_OH_ALLOC: number = undefined;
  OPEN_CON_UI: number = undefined;
  OPEN_CON_UI_ALLOC: number = undefined;
  OPEN_PEN: number = undefined;
  OPEN_PEN_ALLOC: number = undefined;
  OPEN_PEN_ANN: number = undefined;
  OPEN_PEN_ANN_ALLOC: number = undefined;
  OPEN_PEN_OH: number = undefined;
  OPEN_PEN_OH_ALLOC: number = undefined;
  OPEN_PEN_UI: number = undefined;
  OPEN_PEN_UI_ALLOC: number = undefined;
  P_ESTACCMAPPINGID: string;
  REJECTED_OH: number = undefined;
  TOTALNONSAUDIWORKERS: number = undefined;
  TOTALSAUDIANNANDUIONLY: number = undefined;
  TOTALSAUDISANNONLY: number = undefined;
  TOTALSAUDISOHONLY: number = undefined;
  TOTALSAUDISWAGESANNANDUIONLY: number = undefined;
  TOTALSAUDISWAGESANNONLY: number = undefined;
  TOTALSAUDISWAGESOHONLY: number = undefined;
  TOTALSAUDIWORKERS: number = undefined;
  TOTALWAGEOFNONSAUDIWORKERS: number = undefined;
  TOTALWAGEOFSAUDIWORKERS: number = undefined;
  TOTALWAGES: number = undefined;
  TOTALWORKER: number = undefined;
  TOTAL_ADJ_CREDIT: number = undefined;
  TOTAL_CLOSE_ANN: number = undefined;
  TOTAL_CLOSE_DEBIT: number = undefined;
  TOTAL_CLOSE_OH: number = undefined;
  TOTAL_CLOSE_UI: number = undefined;
  TOTAL_CREDIT_ADJ: number = undefined;
  TOTAL_CURRNT_ANN: number = undefined;
  TOTAL_CURRNT_DEBIT: number = undefined;
  TOTAL_CURRNT_OH: number = undefined;
  TOTAL_CURRNT_UI: number = undefined;
  TOTAL_FREE_PENALTY: number = undefined;
  TOTAL_OPEN_ANN: number = undefined;
  TOTAL_OPEN_DEBIT: number = undefined;
  TOTAL_OPEN_OH: number = undefined;
  TOTAL_OPEN_UI: number = undefined;
  TOTALRECEIPTS: number = undefined;
  TRANSFERCREDIT: number = undefined;
  TRANSFERDEBIT: number = undefined;
  VIOLATIONS: number = undefined;
  OHPENAMOUNT: number = undefined;
  UIPENAMOUNT: number = undefined;
  ANNPENAMOUNT: number = undefined;
  VIOLATION_DEDUCTION_REFUND_CREDIT: number = undefined;
  OHCONAMOUNT_CREDIT_ADJ: number = undefined;
  UICONAMOUNT_CREDIT_ADJ: number = undefined;
  ANNCONAMOUNT_CREDIT_ADJ: number = undefined;
  TOTAL_CON_CREDIT_ADJ: number = undefined;
}